requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
customtkinter
PySide6
requests
pyqtgraph
reportlab